Move privacy escaping to a model concern
Using a helper here seems wrong.
