+++ app/helpers/application_helper.rb
@@ -18,29 +18,6 @@ module ApplicationHelper
     recaptcha_tags(display: { theme: 'white' }, ajax: true, hl: I18n.locale)
   end
 
-  def escape_privacy_data(text)
-    return unless text
-
-    text = text.gsub(/([\._a-zA-Z0-9-]+@[\._a-zA-Z0-9-]+)/, ' ')
-    text = text.gsub(/([9|6])+([0-9\s*]{8,})/, ' ')
-    text = text.gsub(/whatsapp/, ' ')
-    text = text.gsub(/whatsupp/, ' ')
-    text = text.gsub(/whatsap/, ' ')
-    text = text.gsub(/whatsap/, ' ')
-    text = text.gsub(/watsap/, ' ')
-    text = text.gsub(/guasap/, ' ')
-    text = text.gsub(/wuassap/, ' ')
-    text = text.gsub(/wuasap/, ' ')
-    text = text.gsub(/wassap/, ' ')
-    text = text.gsub(/wasap/, ' ')
-    text = text.gsub(/guassapp/, ' ')
-    text = text.gsub(/guassap/, ' ')
-    text = text.gsub(/guasapp/, ' ')
-    text = text.gsub(/guasap/, ' ')
-    text = text.gsub(/guasp/, ' ')
-    text
-  end
-
   def errors_for(object)
     errs = object.errors
     return unless errs.any?
